requirements.txt
beautifulsoup4==4.9.3
requests==2.25.0
grpcio-tools==1.33.2